air_pollution/air_pollution_be/serializers/__init__.py
from rest_framework import serializers

class ChatBotRequest(serializers.Serializer):
    question = serializers.CharField(max_length = 1000)

class ChatBotResponse(serializers.Serializer):
    response = serializers.CharField()